+++ packages/components/model-blogpost/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.10](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/model-blogpost@0.0.9...@pubsweet/model-blogpost@0.0.10) (2019-03-06)
+
+**Note:** Version bump only for package @pubsweet/model-blogpost
+
+
+
+
+
+## [0.0.9](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/model-blogpost@0.0.8...@pubsweet/model-blogpost@0.0.9) (2019-03-05)
+
+**Note:** Version bump only for package @pubsweet/model-blogpost
+
+
+
+
+
 ## [0.0.8](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/model-blogpost@0.0.7...@pubsweet/model-blogpost@0.0.8) (2019-02-19)
 
 **Note:** Version bump only for package @pubsweet/model-blogpost