+++ packages/components/xpub-manuscript/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@1.0.3...pubsweet-component-xpub-manuscript@1.0.4) (2019-03-06)
+
+**Note:** Version bump only for package pubsweet-component-xpub-manuscript
+
+
+
+
+
+## [1.0.3](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@1.0.2...pubsweet-component-xpub-manuscript@1.0.3) (2019-03-05)
+
+**Note:** Version bump only for package pubsweet-component-xpub-manuscript
+
+
+
+
+
 ## [1.0.2](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@1.0.1...pubsweet-component-xpub-manuscript@1.0.2) (2019-02-19)
 
 **Note:** Version bump only for package pubsweet-component-xpub-manuscript