+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.17](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/logger@0.2.16...@pubsweet/logger@0.2.17) (2019-03-06)
+
+**Note:** Version bump only for package @pubsweet/logger
+
+
+
+
+
+## [0.2.16](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/logger@0.2.15...@pubsweet/logger@0.2.16) (2019-03-05)
+
+**Note:** Version bump only for package @pubsweet/logger
+
+
+
+
+
 ## [0.2.15](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/logger@0.2.14...@pubsweet/logger@0.2.15) (2019-02-19)
 
 **Note:** Version bump only for package @pubsweet/logger