+++ packages/ui-toolkit/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui-toolkit@2.1.3...@pubsweet/ui-toolkit@2.1.4) (2019-03-06)
+
+**Note:** Version bump only for package @pubsweet/ui-toolkit
+
+
+
+
+
+## [2.1.3](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui-toolkit@2.1.2...@pubsweet/ui-toolkit@2.1.3) (2019-03-05)
+
+**Note:** Version bump only for package @pubsweet/ui-toolkit
+
+
+
+
+
 ## [2.1.2](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui-toolkit@2.1.1...@pubsweet/ui-toolkit@2.1.2) (2019-02-19)
 
 **Note:** Version bump only for package @pubsweet/ui-toolkit